@@ -18,14 +18,14 @@
         }
     ],
     "require": {
-        "php": ">=5.6"
+        "php": ">=7.0"
     },
     "require-dev": {
-        "phpunit/phpunit": ">=5.7"
+        "phpunit/phpunit": ">=7.0"
     },
     "autoload": {
         "psr-4": {
-            "": "src"
+            "Jobs\\ZohoSDK\\": "src/"
         },
         "files": ["src/index.php"]
     }
