@@ -1,7 +1,7 @@
 <?php
 
+namespace Jobs\ZohoSDK\Test\zoho\crm\library\api;
+
 class RelatedListAPIHandlerTest
 {
-	
 }
-?>
