@@ -1,43 +1,43 @@
 <?php
 
-require_once realpath(dirname(__FILE__)."/../../common/APIConstants.php");
+namespace Jobs\ZohoSDK\com\zoho\crm\library\api\response;
+
+use Jobs\ZohoSDK\com\zoho\crm\library\common\APIConstants;
 
 class EntityResponse
 {
-	private $status=null;
-	private $message=null;
-	private $code=null;
-	private $responseJSON=null;
-	private $data=null;
-	private $upsertDetails=array();
-	private $details=null;
-	
-	public function __construct($entityResponseJSON)
-	{
-		$this->responseJSON=$entityResponseJSON;
-		$this->status=$entityResponseJSON[APIConstants::STATUS];
-		$this->message=$entityResponseJSON[APIConstants::MESSAGE];
-		$this->code=$entityResponseJSON[APIConstants::CODE];
-		
-		if(array_key_exists(APIConstants::ACTION,$entityResponseJSON))
-		{
-			$this->upsertDetails[APIConstants::ACTION]=$entityResponseJSON[APIConstants::ACTION];
-		}
-		if(array_key_exists(APIConstants::DUPLICATE_FIELD,$entityResponseJSON))
-		{
-			$this->upsertDetails[APIConstants::DUPLICATE_FIELD]=$entityResponseJSON[APIConstants::DUPLICATE_FIELD];
-		}
-		if(array_key_exists("details",$entityResponseJSON))
-		{
-			$this->details=$entityResponseJSON["details"];
-		}
-	}
+    private $status=null;
+    private $message=null;
+    private $code=null;
+    private $responseJSON=null;
+    private $data=null;
+    private $upsertDetails=[];
+    private $details=null;
+    
+    public function __construct($entityResponseJSON)
+    {
+        $this->responseJSON=$entityResponseJSON;
+        $this->status=$entityResponseJSON[APIConstants::STATUS];
+        $this->message=$entityResponseJSON[APIConstants::MESSAGE];
+        $this->code=$entityResponseJSON[APIConstants::CODE];
+        
+        if (array_key_exists(APIConstants::ACTION, $entityResponseJSON)) {
+            $this->upsertDetails[APIConstants::ACTION]=$entityResponseJSON[APIConstants::ACTION];
+        }
+        if (array_key_exists(APIConstants::DUPLICATE_FIELD, $entityResponseJSON)) {
+            $this->upsertDetails[APIConstants::DUPLICATE_FIELD]=$entityResponseJSON[APIConstants::DUPLICATE_FIELD];
+        }
+        if (array_key_exists('details', $entityResponseJSON)) {
+            $this->details=$entityResponseJSON['details'];
+        }
+    }
 
     /**
      * status
      * @return String
      */
-    public function getStatus(){
+    public function getStatus()
+    {
         return $this->status;
     }
 
@@ -45,7 +45,8 @@ class EntityResponse
      * status
      * @param String $status
      */
-    public function setStatus($status){
+    public function setStatus($status)
+    {
         $this->status = $status;
     }
 
@@ -53,7 +54,8 @@ class EntityResponse
      * message
      * @return String
      */
-    public function getMessage(){
+    public function getMessage()
+    {
         return $this->message;
     }
 
@@ -61,7 +63,8 @@ class EntityResponse
      * message
      * @param String $message
      */
-    public function setMessage($message){
+    public function setMessage($message)
+    {
         $this->message = $message;
     }
 
@@ -69,7 +72,8 @@ class EntityResponse
      * code
      * @return String
      */
-    public function getCode(){
+    public function getCode()
+    {
         return $this->code;
     }
 
@@ -77,7 +81,8 @@ class EntityResponse
      * code
      * @param String $code
      */
-    public function setCode($code){
+    public function setCode($code)
+    {
         $this->code = $code;
     }
 
@@ -86,7 +91,8 @@ class EntityResponse
      * responseJson
      * @return JSONObject
      */
-    public function getResponseJSON(){
+    public function getResponseJSON()
+    {
         return $this->responseJSON;
     }
 
@@ -95,7 +101,8 @@ class EntityResponse
      * data
      * @return Entity Data
      */
-    public function getData(){
+    public function getData()
+    {
         return $this->data;
     }
 
@@ -103,7 +110,8 @@ class EntityResponse
      * data
      * @param Entity Data $data
      */
-    public function setData($data){
+    public function setData($data)
+    {
         $this->data = $data;
     }
     /*
@@ -111,18 +119,16 @@ class EntityResponse
      */
     public function getUpsertDetails()
     {
-    	return $this->upsertDetails;
+        return $this->upsertDetails;
     }
     
     public function setDetails($details)
     {
-    	$this->details=$details;
+        $this->details=$details;
     }
     
     public function getDetails()
     {
-    	return $this->details;
+        return $this->details;
     }
-
 }
-?>
