@@ -1,76 +1,75 @@
 <?php
-require_once realpath(dirname(__FILE__)."/../../common/APIConstants.php");
-require_once realpath(dirname(__FILE__)."/../../exception/ZCRMException.php");
+
+namespace Jobs\ZohoSDK\com\zoho\crm\library\api\response;
+
+use Jobs\ZohoSDK\com\zoho\crm\library\common\APIConstants;
+use Jobs\ZohoSDK\com\zoho\crm\library\exception\APIExceptionHandler;
+use Jobs\ZohoSDK\com\zoho\crm\library\exception\ZCRMException;
 
 class FileAPIResponse
 {
-	private $response=null;
-	private $responseJSON=null;
-	private $httpStatusCode=null;
-	private $responseHeaders=null;
-	private $code=null;
-	private $message=null;
-	private $details=null;
-	private $status=null;
-	
-	public function setFileContent($httpResponse,$httpStatusCode)
-	{
-		$this->httpStatusCode=$httpStatusCode;
-		if($httpStatusCode==APIConstants::RESPONSECODE_NO_CONTENT)
-		{
-			$this->responseJSON=array();
-			$this->responseHeaders=array();
-			$exception=new ZCRMException(APIConstants::INVALID_ID_MSG,$httpStatusCode);
-			$exception->setExceptionCode("No Content");
-			throw $exception;
-		}
-		list($headers, $content) = explode("\r\n\r\n",$httpResponse,2);
-		$headerArray=(explode("\r\n",$headers,50));
-		$headerMap=array();
-		foreach ($headerArray as $key)
-		{
-			if(strpos($key,":")!=false)
-			{
-				$splitArray=explode(":",$key);
-				$headerMap[$splitArray[0]]=$splitArray[1];
-			}
-		}
-		if(in_array($httpStatusCode,APIExceptionHandler::getFaultyResponseCodes()))
-		{
-			$content=json_decode($content,true);
-			$this->responseJSON=$content;
-			$exception=new ZCRMException($content['message'],$httpStatusCode);
-			$exception->setExceptionCode($content['code']);
-			$exception->setExceptionDetails($content['details']);
-			throw $exception;
-		}
-		else if($httpStatusCode==APIConstants::RESPONSECODE_OK)
-		{
-			$this->response=$content;
-			$this->responseJSON=array();
-			$this->status=APIConstants::STATUS_SUCCESS;
-		}
-		$this->responseHeaders=$headerMap;
-		return $this;
-	}
-	
-	public function getFileName()
-	{
-		$contentDisp=self::getResponseHeaders()['Content-Disposition'];
-		$fileName=substr($contentDisp,strrpos($contentDisp,"'")+1,strlen($contentDisp));
-		return $fileName;
-	}
-	
-	public function getFileContent()
-	{
-		return $this->response;
-	}
+    private $response=null;
+    private $responseJSON=null;
+    private $httpStatusCode=null;
+    private $responseHeaders=null;
+    private $code=null;
+    private $message=null;
+    private $details=null;
+    private $status=null;
+    
+    public function setFileContent($httpResponse, $httpStatusCode)
+    {
+        $this->httpStatusCode=$httpStatusCode;
+        if ($httpStatusCode==APIConstants::RESPONSECODE_NO_CONTENT) {
+            $this->responseJSON=[];
+            $this->responseHeaders=[];
+            $exception=new ZCRMException(APIConstants::INVALID_ID_MSG, $httpStatusCode);
+            $exception->setExceptionCode('No Content');
+            throw $exception;
+        }
+        list($headers, $content) = explode("\r\n\r\n", $httpResponse, 2);
+        $headerArray=(explode("\r\n", $headers, 50));
+        $headerMap=[];
+        foreach ($headerArray as $key) {
+            if (strpos($key, ':')!=false) {
+                $splitArray=explode(':', $key);
+                $headerMap[$splitArray[0]]=$splitArray[1];
+            }
+        }
+        if (in_array($httpStatusCode, APIExceptionHandler::getFaultyResponseCodes())) {
+            $content=json_decode($content, true);
+            $this->responseJSON=$content;
+            $exception=new ZCRMException($content['message'], $httpStatusCode);
+            $exception->setExceptionCode($content['code']);
+            $exception->setExceptionDetails($content['details']);
+            throw $exception;
+        } elseif ($httpStatusCode==APIConstants::RESPONSECODE_OK) {
+            $this->response=$content;
+            $this->responseJSON=[];
+            $this->status=APIConstants::STATUS_SUCCESS;
+        }
+        $this->responseHeaders=$headerMap;
+        return $this;
+    }
+    
+    public function getFileName()
+    {
+        $contentDisp=self::getResponseHeaders()['Content-Disposition'];
+        $fileName=substr($contentDisp, strrpos($contentDisp, "'")+1, strlen($contentDisp));
+        return $fileName;
+    }
+    
+    public function getFileContent()
+    {
+        return $this->response;
+    }
 
     /**
      * response
      * @return String
      */
-    public function getResponse(){
+    public function getResponse()
+    {
         return $this->response;
     }
 
@@ -78,7 +77,8 @@ class FileAPIResponse
      * response
      * @param String $response
      */
-    public function setResponse($response){
+    public function setResponse($response)
+    {
         $this->response = $response;
     }
 
@@ -86,7 +86,8 @@ class FileAPIResponse
      * responseJSON
      * @return Array
      */
-    public function getResponseJSON(){
+    public function getResponseJSON()
+    {
         return $this->responseJSON;
     }
 
@@ -94,7 +95,8 @@ class FileAPIResponse
      * responseJSON
      * @param Array $responseJSON
      */
-    public function setResponseJSON($responseJSON){
+    public function setResponseJSON($responseJSON)
+    {
         $this->responseJSON = $responseJSON;
     }
 
@@ -102,7 +104,8 @@ class FileAPIResponse
      * httpStatusCode
      * @return String
      */
-    public function getHttpStatusCode(){
+    public function getHttpStatusCode()
+    {
         return $this->httpStatusCode;
     }
 
@@ -110,7 +113,8 @@ class FileAPIResponse
      * httpStatusCode
      * @param String $httpStatusCode
      */
-    public function setHttpStatusCode($httpStatusCode){
+    public function setHttpStatusCode($httpStatusCode)
+    {
         $this->httpStatusCode = $httpStatusCode;
     }
 
@@ -118,7 +122,8 @@ class FileAPIResponse
      * responseHeaders
      * @return Array
      */
-    public function getResponseHeaders(){
+    public function getResponseHeaders()
+    {
         return $this->responseHeaders;
     }
 
@@ -126,7 +131,8 @@ class FileAPIResponse
      * responseHeaders
      * @param Array $responseHeaders
      */
-    public function setResponseHeaders($responseHeaders){
+    public function setResponseHeaders($responseHeaders)
+    {
         $this->responseHeaders = $responseHeaders;
     }
 
@@ -134,7 +140,8 @@ class FileAPIResponse
      * code
      * @return String
      */
-    public function getCode(){
+    public function getCode()
+    {
         return $this->code;
     }
 
@@ -143,7 +150,8 @@ class FileAPIResponse
      * @param String $code
      * @return NewFileAPIResponse
      */
-    public function setCode($code){
+    public function setCode($code)
+    {
         $this->code = $code;
     }
 
@@ -151,7 +159,8 @@ class FileAPIResponse
      * message
      * @return String
      */
-    public function getMessage(){
+    public function getMessage()
+    {
         return $this->message;
     }
 
@@ -159,7 +168,8 @@ class FileAPIResponse
      * message
      * @param String $message
      */
-    public function setMessage($message){
+    public function setMessage($message)
+    {
         $this->message = $message;
     }
 
@@ -167,7 +177,8 @@ class FileAPIResponse
      * details
      * @return Array
      */
-    public function getDetails(){
+    public function getDetails()
+    {
         return $this->details;
     }
 
@@ -176,14 +187,13 @@ class FileAPIResponse
      * @param Array $details
      * @return NewFileAPIResponse
      */
-    public function setDetails($details){
+    public function setDetails($details)
+    {
         $this->details = $details;
     }
     
     public function getStatus()
     {
-    	return $this->status;
+        return $this->status;
     }
-
 }
-?>
