@@ -1,26 +1,30 @@
 <?php
-require_once realpath(dirname(__FILE__)."/../../common/APIConstants.php");
+
+namespace Jobs\ZohoSDK\com\zoho\crm\library\api\response;
+
+use Jobs\ZohoSDK\com\zoho\crm\library\common\APIConstants;
 
 class ResponseInfo
 {
-	private $moreRecords=null;
-	private $recordCount=null;
-	private $pageNo=null;
-	private $perPage=null;
-	
-	public function __construct($reponseInfoJSON)
-	{
-		$this->moreRecords=(bool)$reponseInfoJSON[APIConstants::MORE_RECORDS];
-		$this->recordCount=$reponseInfoJSON[APIConstants::COUNT]+0;
-		$this->pageNo=$reponseInfoJSON[APIConstants::PAGE]+0;
-		$this->perPage=$reponseInfoJSON[APIConstants::PER_PAGE]+0;
-	}
+    private $moreRecords=null;
+    private $recordCount=null;
+    private $pageNo=null;
+    private $perPage=null;
+    
+    public function __construct($reponseInfoJSON)
+    {
+        $this->moreRecords=(bool)$reponseInfoJSON[APIConstants::MORE_RECORDS];
+        $this->recordCount=$reponseInfoJSON[APIConstants::COUNT]+0;
+        $this->pageNo=$reponseInfoJSON[APIConstants::PAGE]+0;
+        $this->perPage=$reponseInfoJSON[APIConstants::PER_PAGE]+0;
+    }
 
     /**
      * moreRecords
      * @return Boolean
      */
-    public function getMoreRecords(){
+    public function getMoreRecords()
+    {
         return $this->moreRecords;
     }
 
@@ -28,7 +32,8 @@ class ResponseInfo
      * moreRecords
      * @param Boolean $moreRecords
      */
-    public function setMoreRecords($moreRecords){
+    public function setMoreRecords($moreRecords)
+    {
         $this->moreRecords = $moreRecords;
     }
 
@@ -36,7 +41,8 @@ class ResponseInfo
      * recordCount
      * @return int
      */
-    public function getRecordCount(){
+    public function getRecordCount()
+    {
         return $this->recordCount;
     }
 
@@ -44,7 +50,8 @@ class ResponseInfo
      * recordCount
      * @param int $recordCount
      */
-    public function setRecordCount($recordCount){
+    public function setRecordCount($recordCount)
+    {
         $this->recordCount = $recordCount;
     }
 
@@ -52,7 +59,8 @@ class ResponseInfo
      * pageNo
      * @return int
      */
-    public function getPageNo(){
+    public function getPageNo()
+    {
         return $this->pageNo;
     }
 
@@ -60,7 +68,8 @@ class ResponseInfo
      * pageNo
      * @param int $pageNo
      */
-    public function setPageNo($pageNo){
+    public function setPageNo($pageNo)
+    {
         $this->pageNo = $pageNo;
     }
 
@@ -68,7 +77,8 @@ class ResponseInfo
      * perPage
      * @return int
      */
-    public function getPerPage(){
+    public function getPerPage()
+    {
         return $this->perPage;
     }
 
@@ -76,9 +86,8 @@ class ResponseInfo
      * perPage
      * @param int $perPage
      */
-    public function setPerPage($perPage){
+    public function setPerPage($perPage)
+    {
         $this->perPage = $perPage;
     }
-
 }
-?>
