@@ -1,40 +1,43 @@
 <?php
 
+namespace Jobs\ZohoSDK\com\zoho\crm\library\crud;
+
 class ZCRMCustomView
 {
-	private $moduleAPIName=null;
-	private $displayValue=null;
-	private $default=null;
-	private $id=null;
-	private $name=null;
-	private $systemName=null;
-	private $sortBy=null;
-	private $category=null;
-	private $fields=array();
-	private $favorite=null;
-	private $sortOrder=null;
-	private $criteriaPattern=null;
-	private $criteria=null;
-	private $categoriesList=array();
-	private $offLine=null;
-	
-	public function __construct($moduleAPIName,$id)
-	{
-		$this->moduleAPIName=$moduleAPIName;
-		$this->id=$id;
-	}
-	
-	public static function getInstance($moduleAPIName=null,$id)
-	{
-		return new ZCRMCustomView($moduleAPIName, $id);
-	}
-	
+    private $moduleAPIName=null;
+    private $displayValue=null;
+    private $default=null;
+    private $id=null;
+    private $name=null;
+    private $systemName=null;
+    private $sortBy=null;
+    private $category=null;
+    private $fields=[];
+    private $favorite=null;
+    private $sortOrder=null;
+    private $criteriaPattern=null;
+    private $criteria=null;
+    private $categoriesList=[];
+    private $offLine=null;
+    
+    public function __construct($moduleAPIName, $id)
+    {
+        $this->moduleAPIName=$moduleAPIName;
+        $this->id=$id;
+    }
+    
+    public static function getInstance($moduleAPIName=null, $id)
+    {
+        return new ZCRMCustomView($moduleAPIName, $id);
+    }
+    
 
     /**
      * Get the display Name of the customView
      * @return unkown
      */
-    public function getDisplayValue(){
+    public function getDisplayValue()
+    {
         return $this->displayValue;
     }
 
@@ -42,7 +45,8 @@ class ZCRMCustomView
      * Set the display Name of the customView
      * @param unkown $displayValue
      */
-    public function setDisplayValue($displayValue){
+    public function setDisplayValue($displayValue)
+    {
         $this->displayValue = $displayValue;
     }
 
@@ -50,7 +54,8 @@ class ZCRMCustomView
      * Method to know whether the custom view is default one or not
      * @return Boolean value
      */
-    public function isDefault(){
+    public function isDefault()
+    {
         return $this->default;
     }
 
@@ -58,7 +63,8 @@ class ZCRMCustomView
      * Method to set the custom view as default
      * @param unkown $default
      */
-    public function setDefault($default){
+    public function setDefault($default)
+    {
         $this->default = $default;
     }
 
@@ -66,7 +72,8 @@ class ZCRMCustomView
      * Get the customview Id
      * @return customViewId
      */
-    public function getId(){
+    public function getId()
+    {
         return $this->id;
     }
 
@@ -74,14 +81,16 @@ class ZCRMCustomView
      * Set the customview Id
      * @param  $id
      */
-    public function setId($id){
+    public function setId($id)
+    {
         $this->id = $id;
     }
 
     /**
      * Method to get the customview Name
      */
-    public function getName(){
+    public function getName()
+    {
         return $this->name;
     }
 
@@ -89,7 +98,8 @@ class ZCRMCustomView
      * Method to set the customview Name
      * @param  $name
      */
-    public function setName($name){
+    public function setName($name)
+    {
         $this->name = $name;
     }
 
@@ -97,7 +107,8 @@ class ZCRMCustomView
      * Method to get the customview system Name
      * @return System name
      */
-    public function getSystemName(){
+    public function getSystemName()
+    {
         return $this->systemName;
     }
 
@@ -105,7 +116,8 @@ class ZCRMCustomView
      * Method to set the customview system Name
      * @param  $systemName
      */
-    public function setSystemName($systemName){
+    public function setSystemName($systemName)
+    {
         $this->systemName = $systemName;
     }
 
@@ -113,7 +125,8 @@ class ZCRMCustomView
      * Method to get the customview SortBy field Name
      * @return sortBy field
      */
-    public function getSortBy(){
+    public function getSortBy()
+    {
         return $this->sortBy;
     }
 
@@ -121,7 +134,8 @@ class ZCRMCustomView
      * Method to set the customview SortBy field Name
      * @param  $sortBy
      */
-    public function setSortBy($sortBy){
+    public function setSortBy($sortBy)
+    {
         $this->sortBy = $sortBy;
     }
 
@@ -129,7 +143,8 @@ class ZCRMCustomView
      * Method to get the customview Category
      * @return Category
      */
-    public function getCategory(){
+    public function getCategory()
+    {
         return $this->category;
     }
 
@@ -137,7 +152,8 @@ class ZCRMCustomView
      * Method to set the customview Category
      * @param $category
      */
-    public function setCategory($category){
+    public function setCategory($category)
+    {
         $this->category = $category;
     }
 
@@ -145,7 +161,8 @@ class ZCRMCustomView
      * Method to get the customview Fields
      * @return custom view fields
      */
-    public function getFields(){
+    public function getFields()
+    {
         return $this->fields;
     }
 
@@ -153,7 +170,8 @@ class ZCRMCustomView
      * Method to set the customview Fields
      * @param $fields
      */
-    public function setFields($fields){
+    public function setFields($fields)
+    {
         $this->fields = $fields;
     }
 
@@ -161,7 +179,8 @@ class ZCRMCustomView
      * Method to check whether the custom view is default one or not
      * @return favorite
      */
-    public function isFavorite(){
+    public function isFavorite()
+    {
         return $this->favorite;
     }
 
@@ -169,7 +188,8 @@ class ZCRMCustomView
      * Method to set the custom view as default one or not
      * @param $favorite
      */
-    public function setFavorite($favorite){
+    public function setFavorite($favorite)
+    {
         $this->favorite = $favorite;
     }
 
@@ -177,7 +197,8 @@ class ZCRMCustomView
      * Method to get the custom view sort order
      * @return sortOrder
      */
-    public function getSortOrder(){
+    public function getSortOrder()
+    {
         return $this->sortOrder;
     }
 
@@ -185,7 +206,8 @@ class ZCRMCustomView
      * Method to set the custom view sort order type
      * @param $sortOrder
      */
-    public function setSortOrder($sortOrder){
+    public function setSortOrder($sortOrder)
+    {
         $this->sortOrder = $sortOrder;
     }
 
@@ -193,7 +215,8 @@ class ZCRMCustomView
      * Method to get the custom view criteria pattern
      * @return CriteriaPattern
      */
-    public function getCriteriaPattern(){
+    public function getCriteriaPattern()
+    {
         return $this->criteriaPattern;
     }
 
@@ -201,7 +224,8 @@ class ZCRMCustomView
      * Method to set the custom view criteria pattern
      * @param  $criteriaPattern
      */
-    public function setCriteriaPattern($criteriaPattern){
+    public function setCriteriaPattern($criteriaPattern)
+    {
         $this->criteriaPattern = $criteriaPattern;
     }
     
@@ -209,16 +233,18 @@ class ZCRMCustomView
      * Method to get the custom view criteria
      * @return Criteria
      */
-    public function getCriteria(){
-    	return $this->criteria;
+    public function getCriteria()
+    {
+        return $this->criteria;
     }
     
     /**
      * Method to set the custom view criteria
      * @param  $criteria
      */
-    public function setCriteria($criteria){
-    	$this->criteria = $criteria;
+    public function setCriteria($criteria)
+    {
+        $this->criteria = $criteria;
     }
 
 
@@ -226,13 +252,14 @@ class ZCRMCustomView
      * moduleAPIName
      * @return String
      */
-    public function getModuleAPIName(){
+    public function getModuleAPIName()
+    {
         return $this->moduleAPIName;
     }
     
-    public function getRecords(String $sortByField=null, String $sortOrder=null,$startIndex=1,$endIndex=200)
+    public function getRecords(String $sortByField=null, String $sortOrder=null, $startIndex=1, $endIndex=200)
     {
-    	return ZCRMModule::getInstance($apiName)->getRecords($this->id,$sortByField,$sortOrder,$startIndex,$endIndex);
+        return ZCRMModule::getInstance($apiName)->getRecords($this->id, $sortByField, $sortOrder, $startIndex, $endIndex);
     }
 
 
@@ -240,7 +267,8 @@ class ZCRMCustomView
      * categoriesList
      * @return unkown
      */
-    public function getCategoriesList(){
+    public function getCategoriesList()
+    {
         return $this->categoriesList;
     }
 
@@ -248,20 +276,19 @@ class ZCRMCustomView
      * categoriesList
      * @param unkown $categoriesList
      */
-    public function setCategoriesList($categoriesList){
+    public function setCategoriesList($categoriesList)
+    {
         $this->categoriesList = $categoriesList;
     }
     
     
     public function setOffLine($off_line)
     {
-    	$this->offLine=(boolean)$off_line;
+        $this->offLine=(boolean)$off_line;
     }
     
     public function isOffLine()
     {
-    	return $this->offLine;
+        return $this->offLine;
     }
-
 }
-?>
