@@ -1,29 +1,32 @@
 <?php
 
+namespace Jobs\ZohoSDK\com\zoho\crm\library\crud;
+
 class ZCRMJunctionRecord
 {
-	private $id;
-	private $apiName;
-	private $relatedDetails=array();
-	
-	private function __construct($apiName,$id)
-	{
-		$this->apiName=$apiName;
-		$this->id=$id;
-	}
-	public static function getInstance($apiName,$id)
-	{
-		return new ZCRMJunctionRecord($apiName,$id);
-	}
-	
-	
+    private $id;
+    private $apiName;
+    private $relatedDetails=[];
+    
+    private function __construct($apiName, $id)
+    {
+        $this->apiName=$apiName;
+        $this->id=$id;
+    }
+    public static function getInstance($apiName, $id)
+    {
+        return new ZCRMJunctionRecord($apiName, $id);
+    }
+    
+    
 
     /**
      * id
      * @return Long
      * get the ID of the junction record
      */
-    public function getId(){
+    public function getId()
+    {
         return $this->id;
     }
 
@@ -32,7 +35,8 @@ class ZCRMJunctionRecord
      * @return String
      * get the API name of the junction record
      */
-    public function getApiName(){
+    public function getApiName()
+    {
         return $this->apiName;
     }
 
@@ -41,7 +45,8 @@ class ZCRMJunctionRecord
      * @return Mapping Array
      * returns the related data between the modules
      */
-    public function getRelatedDetails(){
+    public function getRelatedDetails()
+    {
         return $this->relatedDetails;
     }
 
@@ -50,9 +55,8 @@ class ZCRMJunctionRecord
      * @param Mapping Array $relatedDetails
      * set the related data between the modules (for adding the relation)
      */
-    public function setRelatedData($fieldApiName,$value){
+    public function setRelatedData($fieldApiName, $value)
+    {
         $this->relatedDetails[$fieldApiName]=$value;
     }
-
 }
-?>
