@@ -1,27 +1,30 @@
 <?php
 
+namespace Jobs\ZohoSDK\com\zoho\crm\library\crud;
+
 class ZCRMLeadConvertMappingField
 {
-	private $apiName;
-	private $id;
-	private $fieldLabel;
-	private $required;
-	private function __construct($apiName,$id)
-	{
-		$this->apiName=$apiName;
-		$this->id=$id;
-	}
-	
-	public static function getInstance($apiName,$id)
-	{
-		return new ZCRMLeadConvertMappingField($apiName,$id);
-	}
+    private $apiName;
+    private $id;
+    private $fieldLabel;
+    private $required;
+    private function __construct($apiName, $id)
+    {
+        $this->apiName=$apiName;
+        $this->id=$id;
+    }
+    
+    public static function getInstance($apiName, $id)
+    {
+        return new ZCRMLeadConvertMappingField($apiName, $id);
+    }
 
     /**
      * apiName
      * @return String
      */
-    public function getApiName(){
+    public function getApiName()
+    {
         return $this->apiName;
     }
 
@@ -29,7 +32,8 @@ class ZCRMLeadConvertMappingField
      * apiName
      * @param String $apiName
      */
-    public function setApiName($apiName){
+    public function setApiName($apiName)
+    {
         $this->apiName = $apiName;
     }
 
@@ -37,7 +41,8 @@ class ZCRMLeadConvertMappingField
      * id
      * @return Long
      */
-    public function getId(){
+    public function getId()
+    {
         return $this->id;
     }
 
@@ -45,7 +50,8 @@ class ZCRMLeadConvertMappingField
      * id
      * @param Long $id
      */
-    public function setId($id){
+    public function setId($id)
+    {
         $this->id = $id;
     }
 
@@ -53,7 +59,8 @@ class ZCRMLeadConvertMappingField
      * fieldLabel
      * @return String
      */
-    public function getFieldLabel(){
+    public function getFieldLabel()
+    {
         return $this->fieldLabel;
     }
 
@@ -61,7 +68,8 @@ class ZCRMLeadConvertMappingField
      * fieldLabel
      * @param String $fieldLabel
      */
-    public function setFieldLabel($fieldLabel){
+    public function setFieldLabel($fieldLabel)
+    {
         $this->fieldLabel = $fieldLabel;
     }
 
@@ -69,7 +77,8 @@ class ZCRMLeadConvertMappingField
      * visible
      * @return Boolean
      */
-    public function isRequired(){
+    public function isRequired()
+    {
         return $this->required;
     }
 
@@ -77,9 +86,8 @@ class ZCRMLeadConvertMappingField
      * visible
      * @param Boolean $visible
      */
-    public function setRequired($visible){
+    public function setRequired($visible)
+    {
         $this->required = $visible;
     }
-
 }
-?>
