@@ -1,104 +1,102 @@
 <?php
-require_once realpath(dirname(__FILE__).'/../api/handler/RelatedListAPIHandler.php');
+
+namespace Jobs\ZohoSDK\com\zoho\crm\library\crud;
+
+use Jobs\ZohoSDK\com\zoho\crm\library\api\handler\RelatedListAPIHandler;
 
 class ZCRMModuleRelation
 {
-	
-	private $label=null;
-	private $apiName=null;
-	private $id=null;
-	private $parentModuleAPIName=null;
-	private $visible=null;
-	
-	private $parentRecord=null;
-	private $junctionRecord;
-	
-	private function __construct($parentModuleAPINameOrParentRecord,$relatedListAPINameOrJunctionRecord)
-	{
-		if($parentModuleAPINameOrParentRecord instanceof ZCRMRecord)
-		{
-			$this->parentRecord=$parentModuleAPINameOrParentRecord;
-		}
-		else
-		{
-			$this->parentModuleAPIName=$parentModuleAPINameOrParentRecord;
-		}
-		
-		if($relatedListAPINameOrJunctionRecord instanceof ZCRMJunctionRecord)
-		{
-			$this->junctionRecord=$relatedListAPINameOrJunctionRecord;
-		}
-		else {
-			$this->apiName=$relatedListAPINameOrJunctionRecord;
-		}
-	}
-	public static function getInstance($parentModuleAPINameOrParentRecord,$relatedListAPIName)
-	{
-		return new ZCRMModuleRelation($parentModuleAPINameOrParentRecord,$relatedListAPIName);
-	}
-	
-	public function getRecords($sortByField=null,$sortOrder=null,$page=1,$perPage=20)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->getRecords($sortByField,$sortOrder,$page,$perPage);
-	}
-	
-	public function getNotes($sortByField=null,$sortOrder=null,$page=1,$perPage=20)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->getNotes($sortByField,$sortOrder,$page,$perPage);
-	}
-	
-	public function addNote($zcrmNote)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->addNote($zcrmNote);
-	}
-	
-	public function updateNote($zcrmNote)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->updateNote($zcrmNote);
-	}
-	
-	public function deleteNote($zcrmNote)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->deleteNote($zcrmNote);
-	}
-	
-	public function getAttachments($page,$perPage)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->getAttachments($page,$perPage);
-	}
-	
-	public function uploadAttachment($filePath)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->uploadAttachment($filePath);
-	}
-	public function uploadLinkAsAttachment($attachmentUrl)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->uploadLinkAsAttachment($attachmentUrl);
-	}
-	
-	public function downloadAttachment($attachmentId)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->downloadAttachment($attachmentId);
-	}
-	public function deleteAttachment($attachmentId)
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord,$this)->deleteAttachment($attachmentId);
-	}
-	public function addRelation()
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord, $this->junctionRecord)->addRelation();
-	}
-	public function removeRelation()
-	{
-		return RelatedListAPIHandler::getInstance($this->parentRecord, $this->junctionRecord)->removeRelation();
-	}
-	
+    private $label=null;
+    private $apiName=null;
+    private $id=null;
+    private $parentModuleAPIName=null;
+    private $visible=null;
+    
+    private $parentRecord=null;
+    private $junctionRecord;
+    
+    private function __construct($parentModuleAPINameOrParentRecord, $relatedListAPINameOrJunctionRecord)
+    {
+        if ($parentModuleAPINameOrParentRecord instanceof ZCRMRecord) {
+            $this->parentRecord=$parentModuleAPINameOrParentRecord;
+        } else {
+            $this->parentModuleAPIName=$parentModuleAPINameOrParentRecord;
+        }
+        
+        if ($relatedListAPINameOrJunctionRecord instanceof ZCRMJunctionRecord) {
+            $this->junctionRecord=$relatedListAPINameOrJunctionRecord;
+        } else {
+            $this->apiName=$relatedListAPINameOrJunctionRecord;
+        }
+    }
+    public static function getInstance($parentModuleAPINameOrParentRecord, $relatedListAPIName)
+    {
+        return new ZCRMModuleRelation($parentModuleAPINameOrParentRecord, $relatedListAPIName);
+    }
+    
+    public function getRecords($sortByField=null, $sortOrder=null, $page=1, $perPage=20)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->getRecords($sortByField, $sortOrder, $page, $perPage);
+    }
+    
+    public function getNotes($sortByField=null, $sortOrder=null, $page=1, $perPage=20)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->getNotes($sortByField, $sortOrder, $page, $perPage);
+    }
+    
+    public function addNote($zcrmNote)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->addNote($zcrmNote);
+    }
+    
+    public function updateNote($zcrmNote)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->updateNote($zcrmNote);
+    }
+    
+    public function deleteNote($zcrmNote)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->deleteNote($zcrmNote);
+    }
+    
+    public function getAttachments($page, $perPage)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->getAttachments($page, $perPage);
+    }
+    
+    public function uploadAttachment($filePath)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->uploadAttachment($filePath);
+    }
+    public function uploadLinkAsAttachment($attachmentUrl)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->uploadLinkAsAttachment($attachmentUrl);
+    }
+    
+    public function downloadAttachment($attachmentId)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->downloadAttachment($attachmentId);
+    }
+    public function deleteAttachment($attachmentId)
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this)->deleteAttachment($attachmentId);
+    }
+    public function addRelation()
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this->junctionRecord)->addRelation();
+    }
+    public function removeRelation()
+    {
+        return RelatedListAPIHandler::getInstance($this->parentRecord, $this->junctionRecord)->removeRelation();
+    }
+    
 
     /**
      * label
      * @return String
      */
-    public function getLabel(){
+    public function getLabel()
+    {
         return $this->label;
     }
 
@@ -106,7 +104,8 @@ class ZCRMModuleRelation
      * label
      * @param String $label
      */
-    public function setLabel($label){
+    public function setLabel($label)
+    {
         $this->label = $label;
     }
 
@@ -114,7 +113,8 @@ class ZCRMModuleRelation
      * apiName
      * @return String
      */
-    public function getApiName(){
+    public function getApiName()
+    {
         return $this->apiName;
     }
 
@@ -122,7 +122,8 @@ class ZCRMModuleRelation
      * apiName
      * @param String $apiName
      */
-    public function setApiName($apiName){
+    public function setApiName($apiName)
+    {
         $this->apiName = $apiName;
     }
 
@@ -130,7 +131,8 @@ class ZCRMModuleRelation
      * id
      * @return Long
      */
-    public function getId(){
+    public function getId()
+    {
         return $this->id;
     }
 
@@ -138,7 +140,8 @@ class ZCRMModuleRelation
      * id
      * @param Long $id
      */
-    public function setId($id){
+    public function setId($id)
+    {
         $this->id = $id;
     }
 
@@ -146,7 +149,8 @@ class ZCRMModuleRelation
      * parentModuleAPIName
      * @return String
      */
-    public function getParentModuleAPIName(){
+    public function getParentModuleAPIName()
+    {
         return $this->parentModuleAPIName;
     }
 
@@ -154,7 +158,8 @@ class ZCRMModuleRelation
      * parentModuleAPIName
      * @param String $parentModuleAPIName
      */
-    public function setParentModuleAPIName($parentModuleAPIName){
+    public function setParentModuleAPIName($parentModuleAPIName)
+    {
         $this->parentModuleAPIName = $parentModuleAPIName;
     }
 
@@ -162,7 +167,8 @@ class ZCRMModuleRelation
      * visible
      * @return boolean
      */
-    public function getVisible(){
+    public function getVisible()
+    {
         return (boolean)$this->visible;
     }
 
@@ -170,7 +176,8 @@ class ZCRMModuleRelation
      * visible
      * @param boolean $visible
      */
-    public function setVisible($visible){
+    public function setVisible($visible)
+    {
         $this->visible = $visible;
     }
 
@@ -178,7 +185,8 @@ class ZCRMModuleRelation
      * parentRecord
      * @return ZCRMRecord
      */
-    public function getParentRecord(){
+    public function getParentRecord()
+    {
         return $this->parentRecord;
     }
 
@@ -186,9 +194,8 @@ class ZCRMModuleRelation
      * parentRecord
      * @param ZCRMRecord $parentRecord
      */
-    public function setParentRecord($parentRecord){
+    public function setParentRecord($parentRecord)
+    {
         $this->parentRecord = $parentRecord;
     }
-
 }
-?>
