@@ -1,7 +1,10 @@
 <?php
-require_once realpath(dirname(__FILE__).'/../api/handler/EntityAPIHandler.php');
-require_once realpath(dirname(__FILE__).'/../common/APIConstants.php');
-require_once 'ZCRMModuleRelation.php';
+
+namespace Jobs\ZohoSDK\com\zoho\crm\library\crud;
+
+use Jobs\ZohoSDK\com\zoho\crm\library\api\handler\EntityAPIHandler;
+use Jobs\ZohoSDK\com\zoho\crm\library\common\APIConstants;
+use Jobs\ZohoSDK\com\zoho\crm\library\exception\ZCRMException;
 
 /**
  * Provides methods for basic CRUD operations of the record.
@@ -10,50 +13,51 @@ require_once 'ZCRMModuleRelation.php';
  */
 class ZCRMRecord
 {
-	private $entityId=null;
-	private $moduleApiName=null;
-	private $lineItems=array();
-	private $lookupLabel=null;
-	private $owner=null;
-	private $createdBy=null;
-	private $modifiedBy=null;
-	private $createdTime=null;
-	private $modifiedTime=null;
-	
-	private $fieldNameVsValue=array();
-	private $properties = array();
-	private $participants = array();
-	private $priceDetails = array();
-	private $layout=null;
-	private $taxList=array();
-	private $lastActivityTime=null;
-	
-	private function __construct($module,$entityId)
-	{
-		$this->moduleApiName=$module;
-		$this->entityId=$entityId;
-	}
-	
-	public static function getInstance($module,$entityId)
-	{
-		return new ZCRMRecord($module,$entityId);
-	}
+    private $entityId=null;
+    private $moduleApiName=null;
+    private $lineItems=[];
+    private $lookupLabel=null;
+    private $owner=null;
+    private $createdBy=null;
+    private $modifiedBy=null;
+    private $createdTime=null;
+    private $modifiedTime=null;
+    
+    private $fieldNameVsValue=[];
+    private $properties = [];
+    private $participants = [];
+    private $priceDetails = [];
+    private $layout=null;
+    private $taxList=[];
+    private $lastActivityTime=null;
+    
+    private function __construct($module, $entityId)
+    {
+        $this->moduleApiName=$module;
+        $this->entityId=$entityId;
+    }
+    
+    public static function getInstance($module, $entityId)
+    {
+        return new ZCRMRecord($module, $entityId);
+    }
 
-	
-	public function addTax($taxIns)
-	{
-		array_push($this->taxList,$taxIns);
-	}
-	
-	public function getTaxList()
-	{
-		return $this->taxList;
-	}
+    
+    public function addTax($taxIns)
+    {
+        array_push($this->taxList, $taxIns);
+    }
+    
+    public function getTaxList()
+    {
+        return $this->taxList;
+    }
     /**
      * entityId
      * @return Long
      */
-    public function getEntityId(){
+    public function getEntityId()
+    {
         return $this->entityId;
     }
 
@@ -61,7 +65,8 @@ class ZCRMRecord
      * entityId
      * @param Long $entityId
      */
-    public function setEntityId($entityId){
+    public function setEntityId($entityId)
+    {
         $this->entityId = $entityId;
     }
 
@@ -69,7 +74,8 @@ class ZCRMRecord
      * moduleApiName
      * @return String
      */
-    public function getModuleApiName(){
+    public function getModuleApiName()
+    {
         return $this->moduleApiName;
     }
 
@@ -77,7 +83,8 @@ class ZCRMRecord
      * moduleApiName
      * @param String $moduleApiName
      */
-    public function setModuleApiName($moduleApiName){
+    public function setModuleApiName($moduleApiName)
+    {
         $this->moduleApiName = $moduleApiName;
     }
 
@@ -85,7 +92,8 @@ class ZCRMRecord
      * Method to get the field value by api name
      * @return String
      */
-    public function getFieldValue($apiName){
+    public function getFieldValue($apiName)
+    {
         return $this->fieldNameVsValue[$apiName];
     }
 
@@ -93,20 +101,22 @@ class ZCRMRecord
      * Method to set the field value for api name
      * @param $apiName,$value
      */
-    public function setFieldValue($apiName,$value){
+    public function setFieldValue($apiName, $value)
+    {
         $this->fieldNameVsValue[$apiName] = $value;
     }
     
     public function getData()
     {
-    	return $this->fieldNameVsValue;
+        return $this->fieldNameVsValue;
     }
 
     /**
      * lineItems
      * @return Array
      */
-    public function getLineItems(){
+    public function getLineItems()
+    {
         return $this->lineItems;
     }
 
@@ -114,15 +124,17 @@ class ZCRMRecord
      * lineItems
      * @param Array $lineItems
      */
-    public function addLineItem($lineItem){
-        array_push($this->lineItems,$lineItem);
+    public function addLineItem($lineItem)
+    {
+        array_push($this->lineItems, $lineItem);
     }
 
     /**
      * lookupLabel
      * @return String
      */
-    public function getLookupLabel(){
+    public function getLookupLabel()
+    {
         return $this->lookupLabel;
     }
 
@@ -130,7 +142,8 @@ class ZCRMRecord
      * lookupLabel
      * @param String $lookupLabel
      */
-    public function setLookupLabel($lookupLabel){
+    public function setLookupLabel($lookupLabel)
+    {
         $this->lookupLabel = $lookupLabel;
     }
 
@@ -138,7 +151,8 @@ class ZCRMRecord
      * owner
      * @return ZCRMUser
      */
-    public function getOwner(){
+    public function getOwner()
+    {
         return $this->owner;
     }
 
@@ -146,7 +160,8 @@ class ZCRMRecord
      * owner
      * @param ZCRMUser $owner
      */
-    public function setOwner($owner){
+    public function setOwner($owner)
+    {
         $this->owner = $owner;
     }
 
@@ -154,7 +169,8 @@ class ZCRMRecord
      * createdBy
      * @return ZCRMUser
      */
-    public function getCreatedBy(){
+    public function getCreatedBy()
+    {
         return $this->createdBy;
     }
 
@@ -162,7 +178,8 @@ class ZCRMRecord
      * createdBy
      * @param ZCRMUser $createdBy
      */
-    public function setCreatedBy($createdBy){
+    public function setCreatedBy($createdBy)
+    {
         $this->createdBy = $createdBy;
     }
 
@@ -170,7 +187,8 @@ class ZCRMRecord
      * modifiedBy
      * @return ZCRMUser
      */
-    public function getModifiedBy(){
+    public function getModifiedBy()
+    {
         return $this->modifiedBy;
     }
 
@@ -178,7 +196,8 @@ class ZCRMRecord
      * modifiedBy
      * @param ZCRMUser $modifiedBy
      */
-    public function setModifiedBy($modifiedBy){
+    public function setModifiedBy($modifiedBy)
+    {
         $this->modifiedBy = $modifiedBy;
     }
 
@@ -186,7 +205,8 @@ class ZCRMRecord
      * createdTime
      * @return String
      */
-    public function getCreatedTime(){
+    public function getCreatedTime()
+    {
         return $this->createdTime;
     }
 
@@ -194,7 +214,8 @@ class ZCRMRecord
      * createdTime
      * @param String $createdTime
      */
-    public function setCreatedTime($createdTime){
+    public function setCreatedTime($createdTime)
+    {
         $this->createdTime = $createdTime;
     }
 
@@ -202,7 +223,8 @@ class ZCRMRecord
      * modifiedTime
      * @return String
      */
-    public function getModifiedTime(){
+    public function getModifiedTime()
+    {
         return $this->modifiedTime;
     }
 
@@ -210,7 +232,8 @@ class ZCRMRecord
      * modifiedTime
      * @param String $modifiedTime
      */
-    public function setModifiedTime($modifiedTime){
+    public function setModifiedTime($modifiedTime)
+    {
         $this->modifiedTime = $modifiedTime;
     }
     
@@ -221,13 +244,12 @@ class ZCRMRecord
      */
     public function create()
     {
-    	if(self::getEntityId() != null)
-    	{
-    		$exception = new ZCRMException("Entity ID MUST be null for create operation.",APIConstants::RESPONSECODE_BAD_REQUEST);
-    		$exception->setExceptionCode("ID EXIST");
-    		throw $exception;
-    	}
-    	return EntityAPIHandler::getInstance($this)->createRecord();
+        if (self::getEntityId() != null) {
+            $exception = new ZCRMException('Entity ID MUST be null for create operation.', APIConstants::RESPONSECODE_BAD_REQUEST);
+            $exception->setExceptionCode('ID EXIST');
+            throw $exception;
+        }
+        return EntityAPIHandler::getInstance($this)->createRecord();
     }
     /**
      * Returns the API response of the record update.
@@ -236,13 +258,12 @@ class ZCRMRecord
      */
     public function update()
     {
-    	if(self::getEntityId() == null)
-    	{
-    		$exception = new ZCRMException("Entity ID MUST NOT be null for update operation.",APIConstants::RESPONSECODE_BAD_REQUEST);
-    		$exception->setExceptionCode("ID MISSING");
-    		throw $exception;
-    	}
-    	return EntityAPIHandler::getInstance($this)->updateRecord();
+        if (self::getEntityId() == null) {
+            $exception = new ZCRMException('Entity ID MUST NOT be null for update operation.', APIConstants::RESPONSECODE_BAD_REQUEST);
+            $exception->setExceptionCode('ID MISSING');
+            throw $exception;
+        }
+        return EntityAPIHandler::getInstance($this)->updateRecord();
     }
     
     /**
@@ -252,85 +273,81 @@ class ZCRMRecord
      */
     public function delete()
     {
-    	if(self::getEntityId() == null)
-    	{
-    		$exception= new ZCRMException("Entity ID MUST NOT be null for delete operation.",APIConstants::RESPONSECODE_BAD_REQUEST);
-    		$exception->setExceptionCode("ID MISSING");
-    		throw $exception;
-    	}
-    	return EntityAPIHandler::getInstance($this)->deleteRecord();
+        if (self::getEntityId() == null) {
+            $exception= new ZCRMException('Entity ID MUST NOT be null for delete operation.', APIConstants::RESPONSECODE_BAD_REQUEST);
+            $exception->setExceptionCode('ID MISSING');
+            throw $exception;
+        }
+        return EntityAPIHandler::getInstance($this)->deleteRecord();
     }
     
-    public function convert($potentialRecord=null,$assignToUser=null)
+    public function convert($potentialRecord=null, $assignToUser=null)
     {
-    	return EntityAPIHandler::getInstance($this)->convertRecord($potentialRecord, $assignToUser);
+        return EntityAPIHandler::getInstance($this)->convertRecord($potentialRecord, $assignToUser);
     }
     
-    public function getRelatedListRecords($relatedListAPIName,$sortByField=null,$sortOrder=null,$page=1, $perPage=20) 
+    public function getRelatedListRecords($relatedListAPIName, $sortByField=null, $sortOrder=null, $page=1, $perPage=20)
     {
-    	return ZCRMModuleRelation::getInstance($this,$relatedListAPIName)->getRecords($sortByField,$sortOrder,$page,$perPage);
+        return ZCRMModuleRelation::getInstance($this, $relatedListAPIName)->getRecords($sortByField, $sortOrder, $page, $perPage);
     }
     
-    public function getNotes($sortByField=null,$sortOrder=null,$page=1, $perPage=20)
+    public function getNotes($sortByField=null, $sortOrder=null, $page=1, $perPage=20)
     {
-    	return ZCRMModuleRelation::getInstance($this,"Notes")->getNotes($sortByField,$sortOrder,$page,$perPage);
+        return ZCRMModuleRelation::getInstance($this, 'Notes')->getNotes($sortByField, $sortOrder, $page, $perPage);
     }
     
     public function addNote($zcrmNoteIns)
     {
-    	if($zcrmNoteIns->getId()!=null)
-    	{
-    		$exception=new ZCRMException("Note ID MUST be null for creating a note.",APIConstants::RESPONSECODE_BAD_REQUEST);
-    		$exception->setExceptionCode("ID EXIST");
-    		throw $exception;
-    	}
-    	return ZCRMModuleRelation::getInstance($this,"Notes")->addNote($zcrmNoteIns);
+        if ($zcrmNoteIns->getId()!=null) {
+            $exception=new ZCRMException('Note ID MUST be null for creating a note.', APIConstants::RESPONSECODE_BAD_REQUEST);
+            $exception->setExceptionCode('ID EXIST');
+            throw $exception;
+        }
+        return ZCRMModuleRelation::getInstance($this, 'Notes')->addNote($zcrmNoteIns);
     }
     
     public function updateNote($zcrmNoteIns)
     {
-    	if($zcrmNoteIns->getId()==null)
-    	{
-    		$exception=new ZCRMException("Note ID MUST NOT be null for updating a note.",APIConstants::RESPONSECODE_BAD_REQUEST);
-    		$exception->setExceptionCode("ID MISSING");
-    		throw $exception;
-    	}
-    	return ZCRMModuleRelation::getInstance($this,"Notes")->updateNote($zcrmNoteIns);
+        if ($zcrmNoteIns->getId()==null) {
+            $exception=new ZCRMException('Note ID MUST NOT be null for updating a note.', APIConstants::RESPONSECODE_BAD_REQUEST);
+            $exception->setExceptionCode('ID MISSING');
+            throw $exception;
+        }
+        return ZCRMModuleRelation::getInstance($this, 'Notes')->updateNote($zcrmNoteIns);
     }
     
     public function deleteNote($zcrmNoteIns)
     {
-    	if($zcrmNoteIns->getId()==null)
-    	{
-    		$exception=new ZCRMException("Note ID MUST NOT be null for deleting a note.",APIConstants::RESPONSECODE_BAD_REQUEST);
-    		$exception->setExceptionCode("ID MISSING");
-    		throw $exception;
-    	}
-    	return ZCRMModuleRelation::getInstance($this,"Notes")->deleteNote($zcrmNoteIns);
+        if ($zcrmNoteIns->getId()==null) {
+            $exception=new ZCRMException('Note ID MUST NOT be null for deleting a note.', APIConstants::RESPONSECODE_BAD_REQUEST);
+            $exception->setExceptionCode('ID MISSING');
+            throw $exception;
+        }
+        return ZCRMModuleRelation::getInstance($this, 'Notes')->deleteNote($zcrmNoteIns);
     }
     
-    public function getAttachments($page=1,$perPage=20)
+    public function getAttachments($page=1, $perPage=20)
     {
-    	return ZCRMModuleRelation::getInstance($this,"Attachments")->getAttachments($page,$perPage);
+        return ZCRMModuleRelation::getInstance($this, 'Attachments')->getAttachments($page, $perPage);
     }
     
     public function uploadAttachment($filePath)
     {
-    	return ZCRMModuleRelation::getInstance($this,"Attachments")->uploadAttachment($filePath);
+        return ZCRMModuleRelation::getInstance($this, 'Attachments')->uploadAttachment($filePath);
     }
     
     public function uploadLinkAsAttachment($attachmentUrl)
     {
-    	return ZCRMModuleRelation::getInstance($this,"Attachments")->uploadLinkAsAttachment($attachmentUrl);
+        return ZCRMModuleRelation::getInstance($this, 'Attachments')->uploadLinkAsAttachment($attachmentUrl);
     }
     
     public function downloadAttachment($attachmentId)
     {
-    	return ZCRMModuleRelation::getInstance($this,"Attachments")->downloadAttachment($attachmentId);
+        return ZCRMModuleRelation::getInstance($this, 'Attachments')->downloadAttachment($attachmentId);
     }
     public function deleteAttachment($attachmentId)
     {
-    	return ZCRMModuleRelation::getInstance($this,"Attachments")->deleteAttachment($attachmentId);
+        return ZCRMModuleRelation::getInstance($this, 'Attachments')->deleteAttachment($attachmentId);
     }
     /**
      * To upload photo to the record
@@ -338,14 +355,14 @@ class ZCRMRecord
      */
     public function uploadPhoto($filePath)
     {
-    	return EntityAPIHandler::getInstance($this)->uploadPhoto($filePath);
+        return EntityAPIHandler::getInstance($this)->uploadPhoto($filePath);
     }
     /**
      * To Download the photo of the record
      */
     public function downloadPhoto()
     {
-    	return EntityAPIHandler::getInstance($this)->downloadPhoto();
+        return EntityAPIHandler::getInstance($this)->downloadPhoto();
     }
     
     /**
@@ -353,31 +370,33 @@ class ZCRMRecord
      */
     public function deletePhoto()
     {
-    	return EntityAPIHandler::getInstance($this)->deletePhoto();
+        return EntityAPIHandler::getInstance($this)->deletePhoto();
     }
 
     public function addRelation(ZCRMJunctionRecord $junctionRecord)
     {
-    	return ZCRMModuleRelation::getInstance($this, $junctionRecord)->addRelation();
+        return ZCRMModuleRelation::getInstance($this, $junctionRecord)->addRelation();
     }
     public function removeRelation(ZCRMJunctionRecord $junctionRecord)
     {
-    	return ZCRMModuleRelation::getInstance($this, $junctionRecord)->removeRelation();
+        return ZCRMModuleRelation::getInstance($this, $junctionRecord)->removeRelation();
     }
 
     /**
      * properties
      * @return HashMap
      */
-    public function getAllProperties(){
-    	return $this->properties;
+    public function getAllProperties()
+    {
+        return $this->properties;
     }
     
     /**
      * properties
      * @return HashMap
      */
-    public function getProperty($propertyName){
+    public function getProperty($propertyName)
+    {
         return $this->properties[$propertyName];
     }
 
@@ -385,7 +404,8 @@ class ZCRMRecord
      * properties
      * @param HashMap $properties
      */
-    public function setProperty($key,$value){
+    public function setProperty($key, $value)
+    {
         $this->properties[$key]=$value;
     }
 
@@ -393,7 +413,8 @@ class ZCRMRecord
      * participants
      * @return Array
      */
-    public function getParticipants(){
+    public function getParticipants()
+    {
         return $this->participants;
     }
 
@@ -401,15 +422,17 @@ class ZCRMRecord
      * participants
      * @param Array $participants
      */
-    public function addParticipant($participant){
-        array_push($this->participants,$participant);
+    public function addParticipant($participant)
+    {
+        array_push($this->participants, $participant);
     }
 
     /**
      * priceDetails
      * @return Array
      */
-    public function getPriceDetails(){
+    public function getPriceDetails()
+    {
         return $this->priceDetails;
     }
 
@@ -417,7 +440,8 @@ class ZCRMRecord
      * priceDetails
      * @param Array $priceDetails
      */
-    public function addPriceDetail($priceDetail){
+    public function addPriceDetail($priceDetail)
+    {
         array_push($this->priceDetails, $priceDetail);
     }
 
@@ -425,7 +449,8 @@ class ZCRMRecord
      * layout
      * @return ZCRMLayout
      */
-    public function getLayout(){
+    public function getLayout()
+    {
         return $this->layout;
     }
 
@@ -433,7 +458,8 @@ class ZCRMRecord
      * layout
      * @param ZCRMLayout $layout
      */
-    public function setLayout($layout){
+    public function setLayout($layout)
+    {
         $this->layout = $layout;
     }
 
@@ -442,7 +468,8 @@ class ZCRMRecord
      * lastActivityTime
      * @return String
      */
-    public function getLastActivityTime(){
+    public function getLastActivityTime()
+    {
         return $this->lastActivityTime;
     }
 
@@ -450,9 +477,8 @@ class ZCRMRecord
      * lastActivityTime
      * @param String $lastActivityTime
      */
-    public function setLastActivityTime($lastActivityTime){
+    public function setLastActivityTime($lastActivityTime)
+    {
         $this->lastActivityTime = $lastActivityTime;
     }
-
 }
-?>
