@@ -1,26 +1,28 @@
 <?php
 
+namespace Jobs\ZohoSDK\com\zoho\crm\library\crud;
+
 class ZCRMRelatedListProperties
 {
-	private $sortBy=null;
-	private $sortOrder=null;
-	private $fields=null;
-	
-	private function __construct()
-	{
-		
-	}
-	
-	public static function getInstance()
-	{
-		return new ZCRMRelatedListProperties();
-	}
+    private $sortBy=null;
+    private $sortOrder=null;
+    private $fields=null;
+    
+    private function __construct()
+    {
+    }
+    
+    public static function getInstance()
+    {
+        return new ZCRMRelatedListProperties();
+    }
 
     /**
      * sortBy
      * @return unkown
      */
-    public function getSortBy(){
+    public function getSortBy()
+    {
         return $this->sortBy;
     }
 
@@ -29,7 +31,8 @@ class ZCRMRelatedListProperties
      * @param unkown $sortBy
      * @return ZCRMRelatedListProperties
      */
-    public function setSortBy($sortBy){
+    public function setSortBy($sortBy)
+    {
         $this->sortBy = $sortBy;
     }
 
@@ -37,7 +40,8 @@ class ZCRMRelatedListProperties
      * sortOrder
      * @return unkown
      */
-    public function getSortOrder(){
+    public function getSortOrder()
+    {
         return $this->sortOrder;
     }
 
@@ -46,7 +50,8 @@ class ZCRMRelatedListProperties
      * @param unkown $sortOrder
      * @return ZCRMRelatedListProperties
      */
-    public function setSortOrder($sortOrder){
+    public function setSortOrder($sortOrder)
+    {
         $this->sortOrder = $sortOrder;
     }
 
@@ -54,7 +59,8 @@ class ZCRMRelatedListProperties
      * fields
      * @return unkown
      */
-    public function getFields(){
+    public function getFields()
+    {
         return $this->fields;
     }
 
@@ -63,9 +69,8 @@ class ZCRMRelatedListProperties
      * @param unkown $fields
      * @return ZCRMRelatedListProperties
      */
-    public function setFields($fields){
+    public function setFields($fields)
+    {
         $this->fields = $fields;
     }
-
 }
-?>
