@@ -1,33 +1,36 @@
 <?php
 
+namespace Jobs\ZohoSDK\com\zoho\crm\library\crud;
+
 /**
- * 
+ *
  * @author sumanth-3058
  *This class is to maintain the details of tax related to Inventory line item
  */
 class ZCRMTax
 {
-	private $taxName=null;
-	private $percentage=null;
-	private $value=null;
-	
-	private function __construct($taxName)
-	{
-		$this->taxName=$taxName;
-	}
-	
-	public static function getInstance($taxName)
-	{
-		return new ZCRMTax($taxName);
-	}
-	
-	
+    private $taxName=null;
+    private $percentage=null;
+    private $value=null;
+    
+    private function __construct($taxName)
+    {
+        $this->taxName=$taxName;
+    }
+    
+    public static function getInstance($taxName)
+    {
+        return new ZCRMTax($taxName);
+    }
+    
+    
 
     /**
      * taxName
      * @return String
      */
-    public function getTaxName(){
+    public function getTaxName()
+    {
         return $this->taxName;
     }
 
@@ -35,7 +38,8 @@ class ZCRMTax
      * taxName
      * @param String $taxName
      */
-    public function setTaxName($taxName){
+    public function setTaxName($taxName)
+    {
         $this->taxName = $taxName;
     }
 
@@ -43,7 +47,8 @@ class ZCRMTax
      * percentage
      * @return Double
      */
-    public function getPercentage(){
+    public function getPercentage()
+    {
         return $this->percentage;
     }
 
@@ -51,7 +56,8 @@ class ZCRMTax
      * percentage
      * @param Double $percentage
      */
-    public function setPercentage($percentage){
+    public function setPercentage($percentage)
+    {
         $this->percentage = $percentage;
     }
 
@@ -59,7 +65,8 @@ class ZCRMTax
      * value
      * @return Double
      */
-    public function getValue(){
+    public function getValue()
+    {
         return $this->value;
     }
 
@@ -67,9 +74,8 @@ class ZCRMTax
      * value
      * @param Double $value
      */
-    public function setValue($value){
+    public function setValue($value)
+    {
         $this->value = $value;
     }
-
 }
-?>
