@@ -1,30 +1,33 @@
 <?php
 
+namespace Jobs\ZohoSDK\com\zoho\crm\library\setup\users;
+
 class ZCRMRole
 {
-	private $name=null;
-	private $id=null;
-	private $reportingTo=null;
-	private $label=null;
-	private $isAdmin=null;
-	
-	private function __construct($roleId,$roleName)
-	{
-		$this->id=$roleId;
-		$this->name=$roleName;
-	}
-	
-	public static function getInstance($roleId,$roleName)
-	{
-		return new ZCRMRole($roleId,$roleName);
-	}
-	
+    private $name=null;
+    private $id=null;
+    private $reportingTo=null;
+    private $label=null;
+    private $isAdmin=null;
+    
+    private function __construct($roleId, $roleName)
+    {
+        $this->id=$roleId;
+        $this->name=$roleName;
+    }
+    
+    public static function getInstance($roleId, $roleName)
+    {
+        return new ZCRMRole($roleId, $roleName);
+    }
+    
 
     /**
      * get the Name of the Role name
      * @return String
      */
-    public function getName(){
+    public function getName()
+    {
         return $this->name;
     }
 
@@ -32,7 +35,8 @@ class ZCRMRole
      * set the Name of the Role name
      * @param String role name $name
      */
-    public function setName($name){
+    public function setName($name)
+    {
         $this->name = $name;
     }
 
@@ -40,7 +44,8 @@ class ZCRMRole
      * get the Id of the Role
      * @return Long
      */
-    public function getId(){
+    public function getId()
+    {
         return $this->id;
     }
 
@@ -48,7 +53,8 @@ class ZCRMRole
      * Set the Id of the Role
      * @param Long $id
      */
-    public function setId($id){
+    public function setId($id)
+    {
         $this->id = $id;
     }
 
@@ -56,7 +62,8 @@ class ZCRMRole
      * get the Reporting to role
      * @return ZCRMUser
      */
-    public function getReportingTo(){
+    public function getReportingTo()
+    {
         return $this->reportingTo;
     }
 
@@ -64,7 +71,8 @@ class ZCRMRole
      * Set the Reporting to role
      * @param ZCRMUser $reportingTo
      */
-    public function setReportingTo($reportingTo){
+    public function setReportingTo($reportingTo)
+    {
         $this->reportingTo = $reportingTo;
     }
 
@@ -72,7 +80,8 @@ class ZCRMRole
      * get the Role label
      * @return String
      */
-    public function getDisplayLabel(){
+    public function getDisplayLabel()
+    {
         return $this->label;
     }
 
@@ -80,7 +89,8 @@ class ZCRMRole
      * Set the Role label
      * @param String $label
      */
-    public function setDisplayLabel($label){
+    public function setDisplayLabel($label)
+    {
         $this->label = $label;
     }
 
@@ -88,7 +98,8 @@ class ZCRMRole
      * Know whether the role is Admin role or not
      * @return boolean
      */
-    public function isAdminRole(){
+    public function isAdminRole()
+    {
         return $this->isAdmin;
     }
 
@@ -96,9 +107,8 @@ class ZCRMRole
      * Set the role as Admin role
      * @param boolean $isAdmin
      */
-    public function setAdminRole($isAdmin){
+    public function setAdminRole($isAdmin)
+    {
         $this->isAdmin = $isAdmin;
     }
-
 }
-?>
