@@ -1,44 +1,46 @@
 <?php
+
+namespace Jobs\ZohoSDK\com\zoho\oauth\common;
+
 class ZohoOAuthConstants
 {
-	const IAM_URL="accounts_url";
-	const SCOPES="scope";
-	const STATE="state";
-	const STATE_OBTAINING_GRANT_TOKEN="OBTAIN_GRANT_TOKEN";
-	const RESPONSE_TYPE="response_type";
-	const RESPONSE_TYPE_CODE="code";
-	const CLIENT_ID="client_id";
-	const CLIENT_SECRET="client_secret";
-	const REDIRECT_URL="redirect_uri";
-	const ACCESS_TYPE="access_type";
-	const ACCESS_TYPE_OFFLINE="offline";
-	const ACCESS_TYPE_ONLINE="online";
-	const PROMPT="prompt";
-	const PROMPT_CONSENT="consent";
-	const GRANT_TYPE="grant_type";
-	const GRANT_TYPE_AUTH_CODE="authorization_code";
-	
-	const GRANT_TYPE_REFRESH="refresh_token";
-	const CODE="code";
-	const GRANT_TOKEN="grant_token";
-	const ACCESS_TOKEN="access_token";
-	const REFRESH_TOKEN="refresh_token";
-	const EXPIRES_IN = "expires_in";
-	const EXPIRIY_TIME = "expiry_time";
-	const PERSISTENCE_HANDLER_CLASS = "persistence_handler_class";
-	const TOKEN = "token";
-	const DISPATCH_TO = "dispatchTo";
-	const OAUTH_TOKENS_PARAM = "oauth_tokens";
-	
-	const OAUTH_HEADER_PREFIX="Zoho-oauthtoken ";
-	const AUTHORIZATION="Authorization";
-	const REQUEST_METHOD_GET="GET";
-	const REQUEST_METHOD_POST="POST";
-	const SANDBOX = "sandbox";
-	const TOKEN_PERSISTENCE_PATH = "token_persistence_path";
-	const DATABASE_PORT = "db_port";
-	const DATABASE_USERNAME = "db_username";
-	const DATABASE_PASSWORD = "db_password";
-	const RESPONSECODE_OK=200;
+    const IAM_URL='accounts_url';
+    const SCOPES='scope';
+    const STATE='state';
+    const STATE_OBTAINING_GRANT_TOKEN='OBTAIN_GRANT_TOKEN';
+    const RESPONSE_TYPE='response_type';
+    const RESPONSE_TYPE_CODE='code';
+    const CLIENT_ID='client_id';
+    const CLIENT_SECRET='client_secret';
+    const REDIRECT_URL='redirect_uri';
+    const ACCESS_TYPE='access_type';
+    const ACCESS_TYPE_OFFLINE='offline';
+    const ACCESS_TYPE_ONLINE='online';
+    const PROMPT='prompt';
+    const PROMPT_CONSENT='consent';
+    const GRANT_TYPE='grant_type';
+    const GRANT_TYPE_AUTH_CODE='authorization_code';
+    
+    const GRANT_TYPE_REFRESH='refresh_token';
+    const CODE='code';
+    const GRANT_TOKEN='grant_token';
+    const ACCESS_TOKEN='access_token';
+    const REFRESH_TOKEN='refresh_token';
+    const EXPIRES_IN = 'expires_in';
+    const EXPIRIY_TIME = 'expiry_time';
+    const PERSISTENCE_HANDLER_CLASS = 'persistence_handler_class';
+    const TOKEN = 'token';
+    const DISPATCH_TO = 'dispatchTo';
+    const OAUTH_TOKENS_PARAM = 'oauth_tokens';
+    
+    const OAUTH_HEADER_PREFIX='Zoho-oauthtoken ';
+    const AUTHORIZATION='Authorization';
+    const REQUEST_METHOD_GET='GET';
+    const REQUEST_METHOD_POST='POST';
+    const SANDBOX = 'sandbox';
+    const TOKEN_PERSISTENCE_PATH = 'token_persistence_path';
+    const DATABASE_PORT = 'db_port';
+    const DATABASE_USERNAME = 'db_username';
+    const DATABASE_PASSWORD = 'db_password';
+    const RESPONSECODE_OK=200;
 }
-?>
